[MISC][WK] create helper and remove unused type
